adds Literals for jobsettings and model modules
