+++ pyproject.toml
@@ -20,9 +20,6 @@ dependencies = [
     "aind-data-schema==0.36.0",
     "aind-data-schema-models==0.1.7",
     "pydantic-settings>=2.0",
-    "pandas",
-    "numpy",
-    "pytz"
 ]
 
 [project.optional-dependencies]
@@ -46,7 +43,8 @@ all = [
 ]
 
 bergamo = [
-    "scanimage-tiff-reader==1.4.1.4"
+    "scanimage-tiff-reader==1.4.1.4",
+    "numpy",
 ]
 
 bruker = [
@@ -64,6 +62,8 @@ openephys = [
     "np_session",
     "npc_ephys",
     "scipy",
+    "pandas",
+    "numpy",
 ]
 
 [tool.setuptools.packages.find]
