+++ src/aind_metadata_mapper/bergamo/session.py
@@ -11,7 +11,7 @@ from datetime import datetime, timedelta
 from decimal import Decimal
 from enum import Enum
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Optional, Tuple, Union, Literal
 from zoneinfo import ZoneInfo
 
 import numpy as np
@@ -45,87 +45,13 @@ from aind_data_schema.core.session import (
     TriggerType,
 )
 from aind_data_schema_models.units import PowerUnit
-from pydantic import Field
-from pydantic_settings import BaseSettings
 from ScanImageTiffReader import ScanImageTiffReader
 
 from aind_metadata_mapper.core import GenericEtl, JobResponse
-
-
-class JobSettings(BaseSettings):
-    """Data that needs to be input by user. Can be pulled from env vars with
-    BERGAMO prefix or set explicitly."""
-
-    input_source: Path = Field(
-        ..., description="Directory of files that need to be parsed."
-    )
-    output_directory: Optional[Path] = Field(
-        default=None,
-        description=(
-            "Directory where to save the json file to. If None, then json"
-            " contents will be returned in the Response message."
-        ),
-    )
-    # mandatory fields:
-    experimenter_full_name: List[str]
-    subject_id: str
-    imaging_laser_wavelength: int  # user defined
-    fov_imaging_depth: int
-    fov_targeted_structure: str
-    notes: Optional[str]
-
-    # fields with default values
-    mouse_platform_name: str = "Standard Mouse Tube"  # FROM RIG JSON
-    active_mouse_platform: bool = False
-    session_type: str = "BCI"
-    iacuc_protocol: str = "2109"
-    # should match rig json:
-    rig_id: str = "442 Bergamo 2p photostim"
-    behavior_camera_names: List[str] = [
-        "Side Face Camera",
-        "Bottom Face Camera",
-    ]
-    ch1_filter_names: List[str] = [
-        "Green emission filter",
-        "Emission dichroic",
-    ]
-    ch1_detector_name: str = "Green PMT"
-    ch1_daq_name: str = "PXI"
-    ch2_filter_names: List[str] = ["Red emission filter", "Emission dichroic"]
-    ch2_detector_name: str = "Red PMT"
-    ch2_daq_name: str = "PXI"
-    imaging_laser_name: str = "Chameleon Laser"
-
-    photostim_laser_name: str = "Monaco Laser"
-    stimulus_device_names: List[str] = ["speaker", "lickport"]  # FROM RIG JSON
-    photostim_laser_wavelength: int = 1040
-    fov_coordinate_ml: Decimal = Decimal("1.5")
-    fov_coordinate_ap: float = Decimal("1.5")
-    fov_reference: str = "Bregma"
-
-    starting_lickport_position: list[float] = [
-        0,
-        -6,
-        0,
-    ]  # in mm from face of the mouse
-    behavior_task_name: str = "single neuron BCI conditioning"
-    hit_rate_trials_0_10: Optional[float] = None
-    hit_rate_trials_20_40: Optional[float] = None
-    total_hits: Optional[float] = None
-    average_hit_rate: Optional[float] = None
-    trial_num: Optional[float] = None
-    # ZoneInfo object doesn't serialize well, so we can define it as a str
-    timezone: str = "US/Pacific"
-
-    class Config:
-        """Config to set env var prefix to BERGAMO"""
-
-        env_prefix = "BERGAMO_"
+from aind_metadata_mapper.bergamo.models import JobSettings
 
 
 # This class makes it easier to flag which tif files are which expected type
-
-
 class TifFileGroup(str, Enum):
     """Type of stimulation a group of files belongs to"""
 
@@ -137,8 +63,6 @@ class TifFileGroup(str, Enum):
 
 # This class will hold the metadata information pulled from the tif files
 # with minimal parsing.
-
-
 @dataclass(frozen=True)
 class RawImageInfo:
     """Raw metadata from a tif file"""
