+++ src/aind_metadata_mapper/fip/models.py
@@ -1,0 +1,35 @@
+"""Module defining JobSettings for FIP ETL"""
+
+from pydantic_settings import BaseSettings
+from typing import List, Optional, Literal
+from pathlib import Path
+from pydantic import Field
+from datetime import datetime
+
+
+class JobSettings(BaseSettings):
+    """Data that needs to be input by user."""
+
+    job_settings_name: Literal["FIP"] = "FIP"
+    output_directory: Optional[Path] = Field(
+        default=None,
+        description=(
+            "Directory where to save the json file to. If None, then json"
+            " contents will be returned in the Response message."
+        ),
+    )
+
+    string_to_parse: str
+    experimenter_full_name: List[str]
+    session_start_time: datetime
+    notes: str
+    labtracks_id: str
+    iacuc_protocol: str
+    light_source_list: List[dict]
+    detector_list: List[dict]
+    fiber_connections_list: List[dict]
+
+    rig_id: str = "ophys_rig"
+    session_type: str = "Foraging_Photometry"
+    mouse_platform_name: str = "Disc"
+    active_mouse_platform: bool = False
