+++ src/aind_metadata_mapper/mesoscope/session.py
@@ -5,41 +5,16 @@ import json
 import sys
 from datetime import datetime
 from pathlib import Path
-from typing import List, Union
+from typing import Union
 
 import tifffile
 from aind_data_schema.core.session import FieldOfView, Session, Stream
 from aind_data_schema_models.modalities import Modality
 from PIL import Image
 from PIL.TiffTags import TAGS
-from pydantic import Field
-from pydantic_settings import BaseSettings
 
 from aind_metadata_mapper.core import GenericEtl
-
-
-class JobSettings(BaseSettings):
-    """Data to be entered by the user."""
-
-    # TODO: for now this will need to be directly input by the user.
-    #  In the future, the workflow sequencing engine should be able to put
-    #  this in a json or we can extract it from SLIMS
-    input_source: Path
-    behavior_source: Path
-    output_directory: Path
-    session_start_time: datetime
-    session_end_time: datetime
-    subject_id: str
-    project: str
-    iacuc_protocol: str = "2115"
-    magnification: str = "16x"
-    fov_coordinate_ml: float = 1.5
-    fov_coordinate_ap: float = 1.5
-    fov_reference: str = "Bregma"
-    experimenter_full_name: List[str] = Field(
-        ..., title="Full name of the experimenter"
-    )
-    mouse_platform_name: str = "disc"
+from aind_metadata_mapper.mesoscope.models import JobSettings
 
 
 class MesoscopeEtl(GenericEtl[JobSettings]):
